adicionando float action button e ajustando o redirecionamento
