+++ kotlin/dio/2-eletric-car-app/EletricCarApp/app/src/main/java/com/example/eletriccarapp/presentation/ui/MainActivity.kt
@@ -1,22 +1,14 @@
 package com.example.eletriccarapp.presentation.ui
 
-import android.content.Intent
 import android.os.Bundle
 import android.util.Log
-import android.widget.Button
-import android.widget.LinearLayout
 import androidx.appcompat.app.AppCompatActivity
-import androidx.recyclerview.widget.RecyclerView
-import androidx.viewpager.widget.ViewPager
 import androidx.viewpager2.widget.ViewPager2
 import com.example.eletriccarapp.R
-import com.example.eletriccarapp.presentation.data.CarFactory
-import com.example.eletriccarapp.presentation.ui.adapter.CarAdapter
 import com.google.android.material.tabs.TabLayout
 
 class MainActivity : AppCompatActivity() {
-    lateinit var btnRedirect: Button
-    lateinit var lista: RecyclerView
+
     lateinit var tabLayout: TabLayout
     lateinit var viewPager: ViewPager2
 
@@ -25,8 +17,6 @@ class MainActivity : AppCompatActivity() {
         Log.d("Lifecycle", "CREATE")
         setContentView(R.layout.activity_main)
         setupView()
-        setupListerners()
-        setupList()
         setupTabs()
     }
 
@@ -53,21 +43,14 @@ class MainActivity : AppCompatActivity() {
 
     private fun setupView() {
         tabLayout = findViewById(R.id.tab_layout)
-        btnRedirect = findViewById(R.id.btn_redirect) //btn_redirect
-        lista = findViewById(R.id.rv_carros)
         viewPager = findViewById(R.id.view_pager)
     }
 
-    private fun setupList() {
-        val adpater = CarAdapter(CarFactory.list)
-        lista.adapter = adpater
-    }
+    private fun setupTabs() {
+        val tabsAdapter = TabAdapter(this)
+        viewPager.adapter = tabsAdapter
 
-    private fun setupListerners() {
-        btnRedirect.setOnClickListener {
-            startActivity( Intent(this, CalcularAutonomiaActivity::class.java))
-        }
-        tabLayout.addOnTabSelectedListener(object : TabLayout.OnTabSelectedListener{
+        tabLayout.addOnTabSelectedListener(object : TabLayout.OnTabSelectedListener {
             override fun onTabSelected(tab: TabLayout.Tab?) {
                 tab?.let {
                     viewPager.currentItem = it.position
@@ -75,21 +58,20 @@ class MainActivity : AppCompatActivity() {
             }
 
             override fun onTabUnselected(tab: TabLayout.Tab?) {
-                //TODO("Not yet implemented")
+
             }
 
             override fun onTabReselected(tab: TabLayout.Tab?) {
-                //TODO("Not yet implemented")
+
             }
 
 
         })
-
-    }
-
-    private  fun setupTabs() {
-        val tabsAdapter = TabAdapter(this)
-        viewPager.adapter = tabsAdapter
-        //tabLayout
+        viewPager.registerOnPageChangeCallback(object : ViewPager2.OnPageChangeCallback() {
+            override fun onPageSelected(position: Int) {
+                super.onPageSelected(position)
+                tabLayout.getTabAt(position)?.select()
+            }
+        })
     }
 }
