+++ kotlin/dio/2-eletric-car-app/EletricCarApp/app/src/main/res/layout/activity_main.xml
@@ -24,8 +24,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/tv_titulo"
-        >
+        app:layout_constraintTop_toBottomOf="@+id/tv_titulo">
 
         <com.google.android.material.tabs.TabItem
             android:layout_width="match_parent"
@@ -40,89 +39,14 @@
     </com.google.android.material.tabs.TabLayout>
 
 
-
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/view_pager"
         android:layout_width="0dp"
         android:layout_height="0dp"
-        app:layout_constraintTop_toBottomOf="@id/tab_layout"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"/>
-
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_carros"
-        android:visibility="gone"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toTopOf="@id/btn_redirect"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/tab_layout"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-tools:listitem="@layout/car_item"
-        />
-
-
-    <Button
-        android:id="@+id/btn_redirect"
-        android:visibility="gone"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_margin="16dp"
-        android:layout_marginTop="4dp"
-        android:text="Calcular"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/tab_layout" />
 
-        />
-
-    <!-- INFO -->
-
-    <TextView
-        android:id="@+id/tv_car2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="16dp"
-        android:text="@string/car2_name"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/btn_redirect" />
-
-
-    <ImageView
-        android:id="@+id/iv_car2_img"
-        android:layout_width="200dp"
-        android:layout_height="120dp"
-        android:layout_marginStart="16dp"
-        android:src="@drawable/ferrari_f80_2025"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tv_car2"
-
-        />
-
-
-    <TextView
-        android:id="@+id/tv_car3"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="16dp"
-        android:text="@string/car3_name"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/iv_car2_img" />
-
-    <ImageView
-        android:id="@+id/iv_car3_img"
-        android:layout_width="200dp"
-        android:layout_height="120dp"
-        android:layout_marginStart="16dp"
-        android:src="@drawable/ferrari_849_testarossa_2026"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tv_car3"
-
-        />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
